fix(type): adjusted types and spacing

+++ src/main/java/br/com/mercadolivre/bootcampw2/grupo11/socialmeli/entities/User.java
@@ -14,6 +14,7 @@ import javax.persistence.*;
 public class User {
     @Id
     @GeneratedValue
-    private long userId;
+    private Integer userId;
+
     private String userName;
 }
